Content removal rules add

+++ web/premises/views.py
@@ -24,6 +24,8 @@ from premises.forms import (ArgumentCreationForm, PremiseCreationForm,
 from premises.signals import (added_premise_for_premise,
                               added_premise_for_contention, reported_as_fallacy,
                               supported_a_premise)
+from premises.templatetags.premise_tags import check_content_deletion
+from newsfeed.models import Entry
 
 
 class ContentionDetailView(DetailView):
@@ -338,9 +340,15 @@ class ArgumentDeleteView(DetailView):
 
     def post(self, request, slug):
         contention = self.get_object()
-        contention.delete()
-        messages.info(request, u"Argümanınız silindi.")
-        return redirect("home")
+        if check_content_deletion(contention):
+            # remove notification
+            Entry.objects.delete(contention.get_newsfeed_type(), contention.id)
+            contention.delete()
+            messages.info(request, u"Argümanınız silindi.")
+            return redirect("home")
+        else:
+            messages.info(request, u"Argümanınız silinecek durumda değil.")
+            return redirect(contention)
 
     delete = post
